This evening I've been investigating why linking with the shared uClibc library was not working when booting busybox as init, even though it workes when run on my desktop system.  I made  boot floppy with just /dev, a few /etc entries, busybox dynamically linked vs uClibc, and /lib with shared uClibc in it.  And just as several people have reported, it didn't work.
So I ran 'ldd /bin/true' and it showed 'true' is linked with the shared library
loader.  But 'ldd ./busybox' showed it was only linked with uClibc.  Doh!  Of
course it couldn't work!

This patch fixes the gcc wrapper so when we link with the shared uClibc
library, we also always link with the correct shared lib loader.  With this
patch in place, I have now sucessfully booted busybox linked vs the shared
uClibc library.  And it works great!
 -Erik