Patch from Jeffrey Damick:
With glibc if you do a res_init() that forces a reread of the nameservers file,
but in uClibc the res_init will not force a reread if it has already read the
file because then the nameservers value is greater than 0.

So res_init must call close_nameservers() before calling open_nameservers().