This should enabled building of uClibc locale support when using uClibc itself. Sorry I didn't test this before the release.
Please remember that the locale data generation tools are not very robust,
so doing something like disabling 8-bit codeset support is likely to break
things.  As it stands, UTF-8 support is required, but I'm not sure I test
for that.

Also, you will notice a difference in the locale data generated by uClibc
verses glibc.  That's because the bg_BG locale specifies use of grouping
in LC_NUMERIC, but supplies no grouping char.  The uClibc locale code
tests for and works around this (at the moment) by disabling grouping.
But the result is slightly different data which ripples throughout the
rest of the tables.