Arthur Shipkowski, art ! videon-central ! com, writes:
I've noticed a few people have posted over the last year about problems
compiling programs that use vfork when pthreads are involved.  Some
detective work turned up that ptfork.c aliases vfork to fork and then tries
to call the original fork as __libc_fork.  This patch removes the aliasing
when there is no MMU present, and uses the same call semantics to call
__libc_vfork.  I then added a symbol to the m68k vfork.S to allow vfork to
be called as __libc_vfork.

The same bug exists in the uClibc CVS, and with a possible tweak this patch
should go through there as well.

Obviously, all other platforms need __libc_vfork as a workable means to call
vfork in order for this to work for them.

Let me know if there are any problems with this patch.

Art Shipkowski
Videon Central Software Engineer
(814)235-1111 x307