Henrik Grindal Bakken writes:
The patch touches a minor (well, not that minor, but perhaps only
rarely encountered) bug in the powerpc dynamic linker.

The problem is that addi is called in inline assembly, but there is no
restriction on the second argument.  In powerpc assembler, if the
second argument to addi is r0, it is taken as the value 0, not the
contents of r0.  This happened to me, making the stack pointer 0 on
the invocation on the application.

The patch is against 0.9.22, but there didn't seem to be any changes
to the relevant section in 0.9.23.