Justin Cormack <justin@street-vision.com> writes: Here is a small patch to add epoll support to uclibc.
The only thing I am not sure about is what needs to be done (if anything)
to make it work on 2.4/2.2 kernels (are the syscall numbers defined but
give errors when called or not?).

Tested on 2.6, though some other issues prevent more extensive testing at
the minute.

Mike Frysinger:
cleaned up patch and added checks to make sure that the syscall actually
exists, returning ENOSYS in cases where it doesn't.