In Bug 116, dicksnippe writes:
uClibc's mkstemp/mktemp tries to read /dev/urandom (or /dev/random) to 
generate random contents for the .XXXXXX part of its argument.  In a 
chrooted environment /dev/[u]random might not be available.  Thus the 
mkstemp call fails.

Add back in the braindamaged gettimeofday/getpid code, but only as a 
fallback for when reading /dev/[u]random fail for whatever reasons.