Aurelien Jacobs writes:
After our last update a new compilation breakage appeared when compiling
gcc:

/home/aurel/geex/geexbox/build.i386/toolchain/i386-pc-linux-uclibc/sysroot/usr/include/bits/uClibc_stdio.h:346: error: expected initializer before 'attribute_hidden'

Some investigations revealed that the following changeset is responsible
for this breakage:

http://uclibc.org/cgi-bin/viewcvs.cgi/trunk/uClibc/libc/sysdeps/linux/common/bits/uClibc_stdio.h?rev=16801&r1=16793&r2=16801&diff_format=h

Reverting just one hunk of this changeset fixed the problem us.
I'm not sure this is the right fix, but please see attached patch.


Look good to me.