libc/x86_64: use ELF-compatible definitions for assembler
Fixes open/read/write/close breakage with NPTL on x86_64 (due to missing
'.type' declaration).

Done this way because there are already ELF-style definitions in assemebler
code for x86_64 and looks like HAVE_ELF is not defined and used this days.
Or the other way around, it'd be a bit strange for x86_64 not to have ELF.

Signed-off-by: Roman I Khimov <khimov@altell.ru>
Signed-off-by: Austin Foxley <austinf@cetoncorp.com>