config: Fix passing defconfig args
Don't rely on argument-shuffling but use correct arguments in
the first place.

In bug#3589 Thomas Suckow writes:

The modified kconfig defines -D to be the same function as -d. In
addition -d does not take an argument and Config.in should be passed as
a non-option argument. The existing argument order appears to work on
distributions such as Fedora but chokes on Cygwin (newlib?).

Signed-off-by: Bernhard Reutner-Fischer <rep.dot.nop@gmail.com>